requirements.txt
openai
python-dotenv
playwright
beautifulsoup4
lxml
